+++ http/hooks/stringifyJSON.js
@@ -6,6 +6,10 @@ module.exports = {
   before: function (req) {
     var contentType = req.headers[CONTENT_TYPE] || JSON_CONTENT_TYPE
 
+    if (req.body instanceof FormData) {
+      return;
+    }
+
     if (contentType === JSON_CONTENT_TYPE && _.isObject(req.body)) {
       req.body = JSON.stringify(req.body);
       req.headers[CONTENT_TYPE] = JSON_CONTENT_TYPE;