+++ test/server/callingResolverWarnings.js
@@ -26,8 +26,9 @@ describe('calling resolver warnings', function () {
 
     it('should warn you that you should be resolving an instance', function () {
       expect(logger.warn).to.be.calledWith(
-        'Warning: You are calling `bar` on the static instance of the ActionCreators \'Foo\'. ' +
-        'You should resolve the instance for the current context'
+        'Warning: You are calling `bar` on the static instance of the ' +
+        'ActionCreators \'Foo\'. You should resolve the instance for ' +
+        'the current context'
       );
     });
   });
@@ -47,8 +48,9 @@ describe('calling resolver warnings', function () {
 
     it('should warn you that you should be resolving an instance', function () {
       expect(logger.warn).to.be.calledWith(
-        'Warning: You are calling `bar` on the static instance of the Store \'Foo\'. ' +
-        'You should resolve the instance for the current context'
+        'Warning: You are calling `bar` on the static instance of ' +
+        'the Store \'Foo\'. You should resolve the instance for ' +
+        'the current context'
       );
     });
   });
@@ -65,8 +67,9 @@ describe('calling resolver warnings', function () {
 
     it('should warn you that you should be resolving an instance', function () {
       expect(logger.warn).to.be.calledWith(
-        'Warning: You are calling `bar` on the static instance of the StateSource \'Foo\'. ' +
-        'You should resolve the instance for the current context'
+        'Warning: You are calling `bar` on the static instance of ' +
+        'the StateSource \'Foo\'. You should resolve the instance for ' +
+        'the current context'
       );
     });
   });